requirements.txt
scikit-image~=0.18.1
numpy~=1.19.5
torchvision~=0.9.1
torch~=1.8.1
Pillow~=8.2.0
tqdm~=4.60.0
matplotlib~=3.4.1
tensorflow~=2.4.1